// apps/sense-companion/vendor/premid-activities/websites/H/HBO Max/types.ts
/* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at https://mozilla.org/MPL/2.0/.
 *
 * Vendored from https://github.com/PreMiD/Activities
 * commit e4c0de9c04a3d310877343bdc05b04bdd7b42603.
 */
export interface MaxAPIResponse {
  data: MainData
  included: IncludedItem[]
  path?: string
}

export interface MainData {
  id: string
  attributes: MainAttributes
  relationships?: ItemRelationships
}

export interface MainAttributes {
  name?: string
  videoType?: 'MOVIE' | 'EPISODE' | 'SPECIAL' | 'EXTRA'
  episodeNumber?: number
  seasonNumber?: number
  alternateId?: string
}

export interface IncludedItem {
  id: string
  type: string
  attributes: IncludedAttributes
  relationships?: ItemRelationships
}

export interface IncludedAttributes {
  alternateId?: string
  name?: string
  episodeNumber?: number
  seasonNumber?: number
  videoType?: 'MOVIE' | 'EPISODE' | 'SPECIAL' | 'EXTRA'
  src?: string
  kind?: string
}

export interface ItemRelationships {
  images?: {
    data: RelationshipRef[]
  }
  show?: {
    data: RelationshipRef
  }
  season?: {
    data: RelationshipRef
  }
  episodes?: {
    data: RelationshipRef[]
  }
}

export interface RelationshipRef {
  id: string
  type: string
}

export interface TitleInfo {
  name?: string
  details?: string
  state?: string
  largeImageKey?: string
}